requirements.txt
autopep8==1.4.3
cffi==1.11.5
cycler==0.10.0
Cython==0.29.2
easydict==1.9
flake8==3.6.0
jedi==0.13.2
kiwisolver==1.0.1
matplotlib==3.0.2
mccabe==0.6.1
msgpack==0.6.0
numpy==1.15.4
opencv-python==3.4.5.20
parso==0.3.1
Pillow==5.4.1
protobuf==3.6.1
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
scipy==1.2.0
six==1.12.0
tensorboardX==1.6
torch==0.4.1
torchvision==0.2.1
